Setup project resources
