+++ Pinch/Assets.xcassets/AccentColor.colorset/Contents.json
@@ -1,6 +1,10 @@
 {
   "colors" : [
     {
+      "color" : {
+        "platform" : "universal",
+        "reference" : "labelColor"
+      },
       "idiom" : "universal"
     }
   ],
