+++ Pinch/Screen/ContentView.swift
@@ -1,0 +1,21 @@
+//
+//  ContentView.swift
+//  Pinch
+//
+//  Created by Егор Шереметов on 26.01.2022.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        Text("Hello, world!")
+            .padding()
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}
